feat: reoganize section to move community impact
